Change disharge current size

+++ HMI-APP/Core/Inc/Nodes/VCU.h
@@ -14,7 +14,7 @@
 
 /* Exported constants --------------------------------------------------------*/
 #define VCU_TIMEOUT    					 (uint32_t) 5000					// ms
-#define MCU_SPEED_MAX 			        (uint8_t) 150
+#define MCU_SPEED_MAX 			       (uint8_t) 150
 #define MCU_CURRENT_MAX     	     (uint8_t) 220
 
 /* Exported struct -----------------------------------------------------------*/
@@ -23,7 +23,7 @@ typedef struct {
 } bms_data_t;
 
 typedef struct {
-	float current;
+	uint8_t current;
 	uint8_t speed;
 } mcu_data_t;
